Address review stage 1 (part 1) of PR 55657
